style: modify Yuv 10bit codeCheck

+++ frameworks/innerkitsimpl/converter/src/image_format_convert_ext_utils.cpp
@@ -383,7 +383,8 @@ static void RGBToYuvP010Param(const RGBDataInfo &rgbInfo, SrcConvertParam &srcPa
         destParam.slice[0] = destInfo.buffer + destInfo.yOffset;
         destParam.slice[1] = destInfo.buffer + destInfo.uvOffset * TWO_SLICES;
     } else {
-        int uvStride = (destParam.width % EVEN_ODD_DIVISOR == 0) ? (destParam.width) : (destParam.width + 1);
+        int uvStride = (destParam.width % EVEN_ODD_DIVISOR == 0) ?
+            static_cast<int>(destParam.width) : static_cast<int>(destParam.width + 1);
         destParam.stride[0] = static_cast<int>(destParam.width);
         destParam.stride[1] = static_cast<int>(uvStride);
         destParam.slice[0] = destInfo.buffer;
@@ -778,9 +779,10 @@ static bool YuvP010ToYuvParam(const YUVDataInfo &yDInfo, SrcConvertParam &srcPar
         destParam.slice[1] = destInfo.buffer + destInfo.uvOffset;
     } else {
         dstyStride = static_cast<int>(destParam.width);
-        dstuvStride = (destParam.width % EVEN_ODD_DIVISOR == 0) ? (destParam.width) : (destParam.width + 1);
+        dstuvStride = (destParam.width % EVEN_ODD_DIVISOR == 0) ?
+            static_cast<int>(destParam.width) : static_cast<int>(destParam.width + 1);
         destParam.slice[0] = destInfo.buffer;
-        destParam.slice[1] = destInfo.buffer + dstyStride * destParam.height;
+        destParam.slice[1] = destInfo.buffer + static_cast<uint32_t>(dstyStride) * destParam.height;
     }
     destParam.stride[0] = dstyStride;
     destParam.stride[1] = dstuvStride;
